Create single factory test
