+++ tests/Functional/Services/WorkerEventFactory/EventHandler/WorkerEventFactoryTest.php
@@ -4,7 +4,11 @@ declare(strict_types=1);
 
 namespace App\Tests\Functional\Services\WorkerEventFactory\EventHandler;
 
+use App\Entity\Job;
+use App\Entity\WorkerEvent;
+use App\Event\EventInterface;
 use App\Services\WorkerEventFactory;
+use App\Tests\AbstractBaseFunctionalTest;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromCompilationFailedEventDataProviderTrait;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromCompilationPassedEventDataProviderTrait;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromCompilationStartedEventDataProviderTrait;
@@ -17,8 +21,9 @@ use App\Tests\DataProvider\WorkerEventFactory\CreateFromJobReadyEventDataProvide
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromJobTimeoutEventDataProviderTrait;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromStepEventDataProviderTrait;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromTestEventDataProviderTrait;
+use webignition\ObjectReflector\ObjectReflector;
 
-class WorkerEventFactoryTest extends AbstractEventHandlerTest
+class WorkerEventFactoryTest extends AbstractBaseFunctionalTest
 {
     use CreateFromStepEventDataProviderTrait;
     use CreateFromTestEventDataProviderTrait;
@@ -33,6 +38,45 @@ class WorkerEventFactoryTest extends AbstractEventHandlerTest
     use CreateFromCompilationPassedEventDataProviderTrait;
     use CreateFromCompilationStartedEventDataProviderTrait;
 
+    private WorkerEventFactory $factory;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $factory = self::getContainer()->get(WorkerEventFactory::class);
+        \assert($factory instanceof WorkerEventFactory);
+        $this->factory = $factory;
+    }
+
+    /**
+     * @dataProvider createDataProvider
+     */
+    public function testCreateForEvent(EventInterface $event, WorkerEvent $expectedWorkerEvent): void
+    {
+        $jobLabel = md5((string) rand());
+        $job = Job::create($jobLabel, '', 600);
+
+        $workerEvent = $this->factory->createForEvent($job, $event);
+
+        $expectedReferenceSource = str_replace('{{ job_label }}', $jobLabel, $expectedWorkerEvent->getReference());
+        ObjectReflector::setProperty(
+            $expectedWorkerEvent,
+            WorkerEvent::class,
+            'reference',
+            md5($expectedReferenceSource)
+        );
+
+        self::assertInstanceOf(WorkerEvent::class, $workerEvent);
+        self::assertNotNull($workerEvent->getId());
+        self::assertSame($expectedWorkerEvent->getType(), $workerEvent->getType());
+        self::assertSame($expectedWorkerEvent->getReference(), $workerEvent->getReference());
+        self::assertSame($expectedWorkerEvent->getPayload(), $workerEvent->getPayload());
+    }
+
+    /**
+     * @return array<mixed>
+     */
     public function createDataProvider(): array
     {
         return array_merge(
@@ -49,12 +93,5 @@ class WorkerEventFactoryTest extends AbstractEventHandlerTest
             $this->createFromCompilationPassedEventDataProvider(),
             $this->createFromCompilationStartedEventDataProvider(),
         );
-    }
-
-    protected function getHandler(): ?WorkerEventFactory
-    {
-        $handler = self::getContainer()->get(WorkerEventFactory::class);
-
-        return $handler instanceof WorkerEventFactory ? $handler : null;
     }
 }
